requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
PyMySQL==0.8.0
requests
google-cloud-storage==1.7.0
oauth2client==4.1.2
six==1.11.0
Flask_Cors==3.0.3
Pillow
sklearn
keras
tensorflow
CherryPy==11.1.0
PasteScript==2.0.2
